# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: zboard API
  version: v0.0.1
  description: |
    zboard single-binary operations console API for subscriber and operator workflows.
    Business APIs use /api/v1. Zero's standard event webhook is exposed at /api/zero/events.
servers:
  - url: http://localhost:8080
  - url: https://api.example.com
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/admin/protocol-endpoints:
    get:
      tags: [nodes]
      summary: List protocol endpoints for administration
      parameters:
        - { name: paged, in: query, schema: { type: boolean } }
        - { name: offset, in: query, schema: { type: integer, minimum: 0, default: 0 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
        - { name: q, in: query, schema: { type: string } }
        - { name: ids, in: query, schema: { type: string }, description: Comma-separated endpoint IDs, at most 100, used to hydrate remote multi-select state. }
        - { name: node_id, in: query, schema: { type: integer, format: int64 } }
        - { name: protocol, in: query, schema: { type: string } }
        - { name: active, in: query, schema: { type: boolean } }
        - { name: deployment_status, in: query, schema: { type: string, enum: [running, succeeded, failed, never] } }
        - { name: sort, in: query, schema: { type: string, enum: [sort_order, id, name, node_id, protocol, multiplier, updated_at] } }
        - { name: direction, in: query, schema: { type: string, enum: [asc, desc] } }
      responses:
        "200":
          description: Protocol endpoint list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - properties:
                      data:
                        oneOf:
                          - type: array
                            items: { $ref: "#/components/schemas/ProtocolEndpoint" }
                          - $ref: "#/components/schemas/ProtocolEndpointPage"
    post:
      tags: [nodes]
      summary: Create a protocol endpoint and queue a full node configuration publish
      description: Persists the encrypted server configuration, then automatically queues validation, atomic activation, restart, status and heartbeat checks for the node's complete current configuration.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProtocolEndpointWriteRequest"
      responses:
        "200": { description: Protocol endpoint configuration saved }
...
  /api/v1/admin/operation-logs:
    get:
      tags: [admin]
      summary: List persisted runtime results for protocol publishes, node kernel operations, and background tasks
      description: Operational output is separate from the security and business audit trail. Results are merged newest first across existing durable operation records.
      parameters:
        - { name: source, in: query, schema: { type: string, enum: [protocol_publish, node_kernel, task] } }
        - { name: status, in: query, schema: { type: string, enum: [queued, running, succeeded, failed] } }
        - { name: node_id, in: query, schema: { type: integer, format: int64 } }
        - { name: protocol_endpoint_id, in: query, schema: { type: integer, format: int64 } }
        - { name: cursor, in: query, schema: { type: string }, description: Opaque bidirectional history cursor returned by page.next_cursor or page.previous_cursor. }
        - { name: from, in: query, schema: { type: string }, description: Inclusive start as YYYY-MM-DD or RFC3339. Defaults to the last 30 UTC days. }
        - { name: to, in: query, schema: { type: string }, description: Inclusive date or exclusive RFC3339 boundary. }
        - { name: offset, in: query, deprecated: true, schema: { type: integer, minimum: 0, default: 0 }, description: Legacy compatibility only; cursor navigation is preferred. }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
      responses:
        "200":
          description: Paginated merged operation records
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - properties:
                      data: { $ref: "#/components/schemas/OperationLogPage" }
        "400": { $ref: "#/components/responses/Error" }
  /api/v1/admin/operation-logs/{source}/{id}:
    get:
      tags: [admin]
      summary: Read one persisted operation result including normalized output and error bodies
      parameters:
        - { name: source, in: path, required: true, schema: { type: string, enum: [protocol_publish, node_kernel, task] } }
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        "200":
          description: Operation detail
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - properties:
                      data: { $ref: "#/components/schemas/OperationLogDetail" }
        "404": { $ref: "#/components/responses/Error" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    NodeConnectorBearer:
      type: http
      scheme: bearer
      bearerFormat: ZeroNodeAPIKey
  responses:
    Error:
      description: Request failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiResponse"
  schemas:
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object
          nullable: true
        error:
          $ref: "#/components/schemas/ApiErrorDetail"
        timestamp:
          type: string
    ApiErrorDetail:
      type: object
      description: Versioned machine-readable detail included on HTTP error responses. The top-level message remains for backward compatibility.
      required: [version, code]
      properties:
        version:
          type: integer
          enum: [1]
        code:
          type: string
          example: validation_failed
        fields:
          type: object
          description: Field-path to localized actionable message. Present when code is validation_failed.
          additionalProperties:
            type: string
    PageMetadata:
      type: object
      required: [offset, limit, total, next_cursor, previous_cursor]
      properties:
        offset: { type: integer, minimum: 0 }
        limit: { type: integer, minimum: 1, maximum: 200 }
        total: { type: integer, format: int64, minimum: 0 }
        next_cursor: { type: string, nullable: true }
        previous_cursor: { type: string, nullable: true }
    PageEnvelope:
      type: object
      required: [page, aggregates, facets]
      properties:
        page: { $ref: "#/components/schemas/PageMetadata" }
        aggregates: { type: object, additionalProperties: true }
        facets: { type: object, additionalProperties: true }
        total: { type: integer, format: int64, minimum: 0, deprecated: true, description: Compatibility alias for page.total. }
        offset: { type: integer, minimum: 0, deprecated: true, description: Compatibility alias for page.offset. }
        limit: { type: integer, minimum: 1, maximum: 200, deprecated: true, description: Compatibility alias for page.limit. }
...
    ProtocolEndpoint:
      type: object
      properties:
        id: { type: integer, format: int64 }
        node_id: { type: integer, format: int64 }
        name: { type: string }
        protocol: { type: string }
        address: { type: string }
        port: { type: integer, minimum: 1, maximum: 65535 }
        public_port: { type: integer, minimum: 1, maximum: 65535 }
        cipher: { type: integer }
        parent_protocol_id: { type: integer, format: int64, nullable: true }
        multiplier_milli: { type: integer, minimum: 1, maximum: 100000 }
        client_config: { type: string }
        optional_config: { type: string, description: JSON object with protocol-specific optional values. }
        tags: { type: string, description: JSON array of presentation and routing tags. }
        is_active: { type: boolean }
    ProtocolDeploymentSummary:
      type: object
      properties:
        id: { type: integer, format: int64 }
        status: { type: string, enum: [running, succeeded, failed] }
        has_error: { type: boolean }
        started_at: { type: string, format: date-time, nullable: true }
        finished_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
    ProtocolEndpointSummary:
      type: object
      required: [id, node_id, name, protocol, address, port, public_port, multiplier_milli, is_active, usage]
      properties:
        id: { type: integer, format: int64 }
        node_id: { type: integer, format: int64 }
        node_name: { type: string }
        name: { type: string }
        protocol: { type: string }
        address: { type: string }
        port: { type: integer, minimum: 1, maximum: 65535 }
        public_port: { type: integer, minimum: 1, maximum: 65535 }
        parent_protocol_id: { type: integer, format: int64, nullable: true }
        multiplier_milli: { type: integer, minimum: 1, maximum: 100000 }
        is_active: { type: boolean }
        sort_order: { type: integer }
        latest_deployment: { $ref: "#/components/schemas/ProtocolDeploymentSummary" }
        usage:
          type: object
          properties:
            active_flows: { type: integer, format: int64, minimum: 0 }
            active_credentials: { type: integer, format: int64, minimum: 0 }
            last_used_at: { type: string, format: date-time, nullable: true }
            used_bytes_today: { type: integer, format: int64, minimum: 0 }
            used_bytes_total: { type: integer, format: int64, minimum: 0 }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    ProtocolEndpointPage:
      allOf:
        - $ref: "#/components/schemas/PageEnvelope"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: "#/components/schemas/ProtocolEndpointSummary" }
...
    OperationLog:
      type: object
      properties:
        id: { type: integer, format: int64 }
        source: { type: string, enum: [protocol_publish, node_kernel, task] }
        action: { type: string }
        status: { type: string, enum: [queued, running, succeeded, failed] }
        target_type: { type: string }
        target_id: { type: integer, format: int64 }
        node_id: { type: integer, format: int64 }
        protocol_endpoint_id: { type: integer, format: int64 }
        summary: { type: string }
        has_output: { type: boolean }
        has_error: { type: boolean }
        started_at: { type: string, format: date-time, nullable: true }
        finished_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
    OperationLogDetail:
      allOf:
        - $ref: "#/components/schemas/OperationLog"
        - type: object
          properties:
            output: { type: string }
            error: { type: string }
    OperationLogPage:
      allOf:
        - $ref: "#/components/schemas/PageEnvelope"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: "#/components/schemas/OperationLog" }
...
    ProtocolEndpointWriteRequest:
      type: object
      required:
        - node_id
        - name
        - protocol
        - address
        - port
        - multiplier_milli
        - config
        - client_config
      properties:
        node_id:
          type: integer
          format: int64
        name:
          type: string
        protocol:
          type: string
          enum: [vmess, vless, trojan, shadowsocks, hysteria2, mieru]
        address:
          type: string
        port:
          type: integer
          minimum: 1
          maximum: 65535
        public_port:
          type: integer
          minimum: 1
          maximum: 65535
        cipher:
          type: integer
        parent_protocol_id:
          type: integer
          format: int64
          nullable: true
        multiplier_milli:
          type: integer
          minimum: 1
          maximum: 100000
        config:
          type: string
          description: Zero inbound protocol JSON object. Its type field must match protocol. The current SSH flow stages this JSON but does not apply it to the Zero runtime.
        client_config:
          type: string
          description: Valid client JSON published through the subscription manifest.
        optional_config:
          type: string
          description: Optional JSON object for protocol-specific settings.
        tags:
          type: string
          description: Optional JSON array of endpoint tags.
        is_active:
          type: boolean
        sort_order:
          type: integer
...